refactor(cofiguration): изменение способа регистрации модуля, поддержка конфигурации в виде передаваемых опций
