--- project/libs/services/src/module/config/common/contracts/index.ts
@@ -1,1 +1,0 @@
-export { CommonConfig } from './common-config';
