+++ project/libs/services/src/module/config/common/index.ts
@@ -1,23 +1,24 @@
 import { registerAs } from '@nestjs/config';
 import { Configuration } from '../configuration';
 import { ConfigNamespace } from '../constants';
-import { CommonConfig } from './contracts';
+import { CommonConfig } from '../contracts';
 import { DEFAULT_PORT } from './constants';
 import { CommonModel } from './common.model';
 
-export const commonConfig = registerAs(
-  ConfigNamespace.Common,
-  (): CommonConfig => {
-    const config = new Configuration({
-      environment: process.env.NODE_ENV as CommonConfig['environment'],
-      port: parseInt(
-        process.env.APPLICATION_PORT ?? DEFAULT_PORT.toString(),
-        10
-      ),
-    });
+export const commonConfig = (options: CommonConfig) => {
+  return registerAs(ConfigNamespace.Common, (): CommonConfig => {
+    const config = new Configuration(
+      options ?? {
+        environment: process.env.NODE_ENV as CommonConfig['environment'],
+        port: parseInt(
+          process.env.APPLICATION_PORT ?? DEFAULT_PORT.toString(),
+          10
+        ),
+      }
+    );
 
     return config.validate(CommonModel);
-  }
-);
+  });
+};
 
 export { CommonConfig };
