+++ project/libs/services/src/module/config/contracts/config-module-options.ts
@@ -1,0 +1,3 @@
+import { CommonConfig, DatabaseConfig } from '.';
+
+export type ConfigModuleOptions = CommonConfig & DatabaseConfig;
