+++ project/libs/services/src/module/config/contracts/database-config.ts
@@ -1,8 +1,8 @@
-export interface DatabaseConfig {
+export type DatabaseConfig = {
   name: string;
   host: string;
   port: number;
   user: string;
   password: string;
   authBase: string;
-}
+};
