+++ project/libs/services/src/module/config/contracts/index.ts
@@ -1,0 +1,3 @@
+export type { CommonConfig } from './common-config';
+export type { DatabaseConfig } from './database-config';
+export type { ConfigModuleOptions } from './config-module-options';
