--- project/libs/services/src/module/config/db/contracts/index.ts
@@ -1,1 +1,0 @@
-export type { DatabaseConfig } from './database-config';
