+++ project/libs/services/src/module/config/db/db.model.ts
@@ -1,5 +1,5 @@
 import { IsNumber, IsString } from 'class-validator';
-import { DatabaseConfig } from './contracts';
+import { DatabaseConfig } from '../contracts';
 import { EnvValidationMessage } from './constants';
 
 export class DatabaseModel implements DatabaseConfig {
