+++ project/libs/services/src/module/config/db/utils/get-mongoose-options.ts
@@ -1,6 +1,6 @@
 import { MongooseModuleAsyncOptions } from '@nestjs/mongoose';
 import { ConfigService } from '@nestjs/config';
-import { DatabaseConfig } from '../contracts';
+import { DatabaseConfig } from '../../contracts';
 import { ConfigNamespace } from '../../constants';
 import { getMongoConnectionString } from './get-mongo-connection-string';
