+++ project/libs/services/src/module/config/index.ts
@@ -1,6 +1,7 @@
 import path from 'path';
-import { Module } from '@nestjs/common';
-import { ConfigModule } from '@nestjs/config';
+import { Module, DynamicModule } from '@nestjs/common';
+import { ConfigModule as ConfigModuleRoot } from '@nestjs/config';
+import { ConfigModuleOptions } from './contracts';
 import { dbConfig } from './db';
 import { commonConfig } from './common';
 
@@ -8,17 +9,25 @@ const envFilePath = `apps/${path.basename(__dirname)}/.env.${
   process.env.NODE_ENV
 }`;
 
-@Module({
-  imports: [
-    ConfigModule.forRoot({
-      isGlobal: true,
-      cache: true,
-      load: [dbConfig, commonConfig],
-      envFilePath: envFilePath,
-    }),
-  ],
-})
-export class ConfigModuleRoot {}
+@Module({})
+export class ConfigModule {
+  static forRoot(options?: ConfigModuleOptions): DynamicModule {
+    const dbConfigRegister = dbConfig(options);
+    const commonConfigRegister = commonConfig(options);
+    return {
+      module: ConfigModule,
+      imports: [
+        ConfigModuleRoot.forRoot({
+          isGlobal: true,
+          cache: true,
+          load: [dbConfigRegister, commonConfigRegister],
+          envFilePath: envFilePath,
+        }),
+      ],
+      exports: [],
+    };
+  }
+}
 
 export * from './constants';
 export * from './db';
